refactor: Abstracted further DriveFileManager DB operations
