+++ .package.resolved
@@ -87,8 +87,7 @@
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/Infomaniak/ios-core",
       "state" : {
-        "revision" : "254b41468a64bc926690ce676d50dc182d4911f3",
-        "version" : "9.0.0"
+        "revision" : "c44c0968a5caef41f006b54fdd928e1da2184416"
       }
     },
     {
